ReadMe.md
With this homework assigment I used bootstrap formulas and traditional CSS and html.

With the nav bar I used the code from bootstrap, and it connected all three html's. "About", "Portfolio"( I have it under Practice.html because I used this as a dumby html but it was responding better than the original) and "Contact".

With the CSS's I updated the nav bar, the weight, Font Family and color or the font. Chose the background color of the background and the color of the container.

My Contact and Index pages passed the validator. But my portfolio page came out with an error for my image lines. It said I had duplicate attribute "class". But if I correct those mistakes, it would mess with the image sizes. I could use some help with that. Sorry, given the time constraint, I didn't want to fall back on Java.